Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -2,7 +2,7 @@
 #include<math.h>
 int main(){
     float r,pi,A;
-   int const pi = 3.14;
+   float const pi = 3.14;
 
     scanf("%f",&r);
     A = pi *r*r;
